+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add ClassMetadata.kind
 * Add ConstructorMetadata.isPrimary
 * Add DeclarationMetadata.visibility
+* Add ParameterMetadata.hasDefaultValue
 * Update to Kotlin 1.2.0
 
 ## 1.0.0 (Feb 15, 2016)